refactoring cleaning
